Modified makesymlinks.sh. Fixed bug when no old rc files found. Nicer prints

+++ makesymlinks.sh
@@ -12,20 +12,23 @@ files="bashrc vimrc"    # list of files/folders to symlink in homedir
 ##########
 
 # create dotfiles_old in homedir
-echo "Creating $olddir for backup of any existing dotfiles in ~"
+echo -ne "Creating $olddir for backup of any existing dotfiles in ~ ..."
 mkdir -p $olddir
-echo "...done"
+echo " done"
 
 # change to the dotfiles directory
-echo "Changing to the $dir directory"
+echo -ne "Changing to the $dir directory ..."
 cd $dir
-echo "...done"
+echo " done"
 
 # move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks from the homedir to any files in the ~/dotfiles directory specified in $files
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file ~/dotfiles_old/
-    echo "Creating symlink to $file in home directory."
+    if [ -f $file ]
+    then
+        echo "Moving old $file from ~ to $olddir"
+        mv ~/.$file ~/dotfiles_old/
+    fi
+    echo "Creating symlink to new $file in home directory."
     ln -s $dir/$file ~/.$file
 done
